@@ -2,7 +2,7 @@ import { Component } from '@angular/core';
 import { HeaderComponent } from './components/header/header.component';
 import { ProductListComponent } from './components/product-list/product-list.component';
 import { CartComponent } from './components/cart/cart.component';
-import { HttpClientModule } from '@angular/common/http';
+import {} from '@angular/common/http';
 
 @Component({
   selector: 'app-root',
@@ -11,7 +11,11 @@ import { HttpClientModule } from '@angular/common/http';
     HeaderComponent, 
     ProductListComponent, 
     CartComponent,
-    HttpClientModule  
+    
+// TODO: `HttpClientModule` should not be imported into a component directly.
+// Please refactor the code to add `provideHttpClient()` call to the provider list in the
+// application bootstrap logic and remove the `HttpClientModule` import from this component.
+HttpClientModule  
   ],
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.scss']
