@@ -3,14 +3,18 @@ import { CommonModule } from '@angular/common';
 import { Product } from '../../models/product.model';
 import { ProductItemComponent } from '../product-item/product-item.component';
 import { ProductService } from '../../services/product.service';
-import { HttpClientModule } from '@angular/common/http';
+import {} from '@angular/common/http';
 
 @Component({
   selector: 'app-product-list',
   standalone: true,
   templateUrl: './product-list.component.html',
   styleUrls: ['./product-list.component.scss'],
-  imports: [CommonModule, HttpClientModule, ProductItemComponent]
+  imports: [CommonModule, 
+// TODO: `HttpClientModule` should not be imported into a component directly.
+// Please refactor the code to add `provideHttpClient()` call to the provider list in the
+// application bootstrap logic and remove the `HttpClientModule` import from this component.
+HttpClientModule, ProductItemComponent]
 })
 
 export class ProductListComponent implements OnInit {
